+++ tests/test_color.cpp
@@ -4,6 +4,12 @@
 #include <libtcod/console.hpp>
 #include <sstream>
 
+/// @brief String output for TCODColor.
+std::ostream& operator<<(std::ostream& os, const TCODColor& color) {
+  return os << "{" << std::to_string(color.r) << ", " << std::to_string(color.g) << ", " << std::to_string(color.b)
+            << "}";
+}
+
 TEST_CASE("Color conversions") {
   {
     const auto tile = TCOD_ConsoleTile{
@@ -58,4 +64,32 @@ TEST_CASE("Color IO") {
     stream >> rgba;
     CHECK(rgba == TCOD_ColorRGBA{1, 2, 3, 4});
   }
+}
+TEST_CASE("TCODColor::genMap") {
+  static constexpr auto map_constexpr = TCODColor::genMap<16>(
+      std::array{tcod::ColorRGB{0, 0, 0}, tcod::ColorRGB{255, 0, 0}, tcod::ColorRGB{255, 255, 255}},
+      std::array{0, 8, 15});
+  CHECK(map_constexpr.at(0) == tcod::ColorRGB{0, 0, 0});
+  CHECK(map_constexpr.at(8) == tcod::ColorRGB{255, 0, 0});
+  CHECK(map_constexpr.at(15) == tcod::ColorRGB{255, 255, 255});
+
+  auto old_map = std::array<TCODColor, 16>{};
+  TCODColor::genMap(
+      old_map.data(),
+      3,
+      std::array{TCODColor{0, 0, 0}, TCODColor{255, 0, 0}, TCODColor{255, 255, 255}}.data(),
+      std::array{0, 8, 15}.data());
+  for (size_t i{0}; i < map_constexpr.size(); ++i) {
+    CHECK(map_constexpr.at(i) == tcod::ColorRGB{old_map.at(i)});
+  }
+}
+TEST_CASE("Color math") {
+  CHECK(TCODColor{0, 1, 200} + TCODColor{0, 2, 200} == TCODColor{0, 3, 255});
+
+  CHECK(TCODColor{100, 100, 100} - TCODColor{0, 1, 200} == TCODColor{100, 99, 0});
+
+  CHECK(TCODColor{128, 128, 128} * TCODColor{0, 128, 255} == TCODColor{0, 64, 128});
+
+  CHECK(TCODColor{0, 10, 100} * 10.0f == TCODColor{0, 100, 255});
+  CHECK(TCODColor{0, 10, 100} * -10.0f == TCODColor{0, 0, 0});
 }