+++ CHANGELOG.md
@@ -8,7 +8,8 @@ Versions since `1.7.0` only track ABI breaks and not API breaks.
 
 ## [Unreleased]
 ### Added
-- New `TCODImage::getSize()` overload which returns a value instead if taking output references.
+- New `TCODImage::getSize()` overload which returns a value instead of taking output references.
+- New `TCODColor::genMap` overload which supports constexpr.
 
 ## [1.23.1] - 2022-11-09
 ### Changed